+++ components/keyemulate/api/src/main/kotlin/com/flipperdevices/keyemulate/api/KeyEmulateApi.kt
@@ -2,9 +2,13 @@ package com.flipperdevices.keyemulate.api
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.flipperdevices.bridge.dao.api.model.FlipperKey
+import com.flipperdevices.keyemulate.model.EmulateConfig
 
 interface KeyEmulateApi {
     @Composable
-    fun ComposableEmulateButton(modifier: Modifier, flipperKey: FlipperKey)
+    fun ComposableEmulateButton(
+        modifier: Modifier,
+        emulateConfig: EmulateConfig,
+        isSynchronized: Boolean
+    )
 }